b3dtile::Tile クラスをいくつかのファイルに分割
